🧬 Pull the logic about loading configuration up to abstract class *TemplateConfigLoadable* to let data model could be more clean and simple.
